Fixedd warning introduced in JsiSkShaderBuilder
Cast from long (size_t) -> int.

+++ package/cpp/api/JsiSkRuntimeShaderBuilder.h
@@ -39,7 +39,7 @@ namespace RNSkia {
                 auto e = jsiValue.getValueAtIndex(runtime, i).asNumber();
                 value.push_back(e);
             }
-            getObject()->uniform(name.c_str()).set(value.data(), size);
+            getObject()->uniform(name.c_str()).set(value.data(), (int)size);
             return jsi::Value::undefined();
         }
